refactor: introduce shared StoryData typed contract
